requirements.txt
streamlit
tensorflow
mysql-connector-python
pandas
matplotlib
numpy
pillow
sqlalchemy